# perf(transport/http/binding): optimize EncodeURL performance for paths without placeholders